requirements.txt
langchain==0.0.27
PyPDF2==3.0.1
python-dotenv==0.21.1
streamlit==1.18.1
faiss-cpu==1.7.4
gpt4all==0.3.5


